fix(project): reorganized svgs under core ui component, footer focus styles tweak
