+++ libs/ui/src/lib/logo/logo.module.scss
@@ -19,6 +19,7 @@
     }
 }
 
+.digdirLogoLink,
 .logoLink {
     text-decoration: none;
     padding: 0.5rem;
